requirements.txt
python-telegram-bot==20.1
Pillow